% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allTypesDemoCircuit}
\alias{allTypesDemoCircuit}
\title{A circuit with every interaction type for simulations}
\format{
A data frame with 6 rows and 3 variables
}
\usage{
allTypesDemoCircuit
}
\description{
This data contains an artificial 4-gene circuit with genes A, B,
C, D which contains every possible type of gene interaction simulated by the
sRACIPE package. The circuit is designed so that each pair of genes has only
one type of interaction and each interaction appears only once, which makes
it useful for testing.
}
\keyword{datasets}
