% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{annotation,RacipeSE-method}
\alias{annotation,RacipeSE-method}
\title{A method to get the annotation}
\usage{
\S4method{annotation}{RacipeSE}(object, ...)
}
\arguments{
\item{object}{RacipeSE object}

\item{...}{Additional arguments, for use in specific methods.}
}
\value{
character
}
\description{
A method to get the annotation
}
\examples{

data("demoCircuit")
rSet <- sRACIPE::sracipeSimulate(circuit = demoCircuit, numModels = 100)
ann <- annotation(rSet)
annotation(rSet) <- ann

}
