% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{annotation<-,RacipeSE,ANY-method}
\alias{annotation<-,RacipeSE,ANY-method}
\title{A method to set the circuit name or annotation}
\usage{
\S4method{annotation}{RacipeSE,ANY}(object, ...) <- value
}
\arguments{
\item{object}{RacipeSE object}

\item{...}{Additional arguments, for use in specific methods.}

\item{value}{annotation character}
}
\value{
A RacipeSE object
}
\description{
A method to set the circuit name or annotation
}
\examples{
data("demoCircuit")
rSet <- sRACIPE::sracipeSimulate(circuit = demoCircuit, numModels = 100)
ann <- annotation(rSet)
annotation(rSet) <- ann
}
