% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarity.R
\name{calcHeatmapSimilarity}
\alias{calcHeatmapSimilarity}
\title{Calculates the similarity between two gene expression data.}
\usage{
calcHeatmapSimilarity(data.reference, data.simulation,
  cluster.cut = NULL, n.clusters = 3, p.value = 0.05, permuted.var,
  permutations = 1000, cor.method = "spearman",
  cluster.method = "ward.D2", method = "pvalue", buffer = 1e-04,
  permutation.method = "simulation", return.data = FALSE)
}
\arguments{
\item{data.reference}{Matrix. The reference data matrix, for example, the experimental gene expression values}

\item{data.simulation}{Matrix. The data matrix to be compared.}

\item{cluster.cut}{(optional) Integer vector. Clsuter numbers assigned to reference data.
If cluster.cut is missing, hierarchical clustering using /code{ward.D2}
and /code{distance  = (1-cor(x, method = "spear"))/2} will be used to cluster the reference data.}

\item{n.clusters}{(optional) Integer. The number of clusters in which the
reference data should be clustered for comparison.
Not needed if cluster.cut is provided.}

\item{p.value}{(optional) Numeric. p-value to consider two gene expression
 sets as belonging to same cluster.
Ward's method with spearman correlation is used to determine if a
model belongs to a specific cluster.}

\item{permuted.var}{(optional) Similarity scores computed after permutations.}

\item{cor.method}{(optional) Correlation method. Default method is "spearman". For single cell data, use "kendall"}

\item{cluster.method}{(optional) Character - default \code{ward.D2}, other
options include \code{complete}. Clustering method to be used to cluster the experimental data.
\code{\link[stats]{hclust}} for other options.}

\item{permutation.method}{"sample" or "reference"}
}
\value{
A list containing the KL distance of new cluster distribution from reference data and
the probability of each cluster in the reference and simulated data.
}
\description{
Comparison is done across columns, i.e., how similar are the columns in the two dataset.
For gene expression data, format data so that gene names are in rows and samples in columns.
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}

