% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{circuit<-}
\alias{circuit<-}
\alias{circuit<-,RacipeSE-method}
\alias{circuit-set}
\title{Initialize the circuit}
\usage{
circuit(.object) <- value

\S4method{circuit}{RacipeSE}(.object) <- value
}
\arguments{
\item{.object}{RacipeSE object}

\item{value}{data.frame containing the circuit information}
}
\value{
data.frame
}
\description{
Initialize the circuit from a topology file
or a \code{data.frame}
A typical topology file looks like
\tabular{lcr}{
  Source \tab Target \tab Type \cr
  geneA \tab geneB \tab 2 \cr
  geneB \tab geneC \tab 1 \cr
  geneB \tab geneA \tab 2
}
Here the regulation type is specified by number - activation: \code{1},
 inhibition: \code{2}
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockDownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}}
}

\examples{
RacipeSet <- RacipeSE()
data("demoCircuit")
circuit(RacipeSet) <- demoCircuit
circuit(RacipeSet)
}
