% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{circuit}
\alias{circuit}
\alias{circuit,RacipeSE-method}
\title{Method to get the circuit}
\usage{
circuit(.object)

\S4method{circuit}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
A dataframe
}
\description{
The circuit file should contain three columns with headers,
"Source" "Target" "Type"
Here "Source" and "Target" are the names of the genes and "Type" refers to
the regulation, "1" if source activates target and "2" if source inhibits
target.
}
\examples{
rs <- RacipeSE()
data("demoCircuit")
circuit(rs) <- demoCircuit
circuitDataFrame <- circuit(rs)

}
