% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{config<-}
\alias{config<-}
\alias{config<-,RacipeSE-method}
\alias{config-set}
\title{A method to access the simulation hyperparameters}
\usage{
config(.object) <- value

\S4method{config}{RacipeSE}(.object) <- value
}
\arguments{
\item{.object}{RacipeSE object}

\item{value}{list. Configuration as a list}
}
\value{
\code{RacipeSE} object
}
\description{
The hyperparameters like number of models, range from which
parameters are to be sampled, simulation time etc.
}
\examples{
rSet <- RacipeSE()
tmpConfig <- config(rSet)
config(rSet) <- tmpConfig
}
