% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{config}
\alias{config}
\alias{config,RacipeSE-method}
\title{A method to access the simulation hyperparameters}
\usage{
config(.object)

\S4method{config}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
list
}
\description{
The hyperparameters like number of models, range from which
parameters are to be sampled, simulation time etc.
}
\examples{
RacipeSet <- RacipeSE()
data("demoCircuit")
circuit(RacipeSet) <- demoCircuit
config(RacipeSet)
}
