% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoupledToggleSwitchSA}
\alias{CoupledToggleSwitchSA}
\title{A coupled toggle switch circuit with self activations}
\format{data frame}
\usage{
CoupledToggleSwitchSA
}
\description{
A coupled toggle switch circuit with self activations
}
\keyword{datasets}
