% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoCircuit}
\alias{demoCircuit}
\title{A toggle switch circuit for demonstrations}
\format{data frame}
\usage{
demoCircuit
}
\description{
A toggle switch circuit for demonstrations
}
\keyword{datasets}
