% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.loadNetworkFile}
\alias{.loadNetworkFile}
\title{Loads the network/topology file.}
\usage{
.loadNetworkFile(networkFile = "inputs/test.net")
}
\arguments{
\item{networkFile}{Network file name}
}
\value{
Network as a dataframe
}
\description{
The network file should contain three columns with headers,
"Source" "Target" "Type"
Here "Source" and "Target" are the names of the genes and "Type" refers to
the regulation, "1" if source activates target and "2" if source inhibits
target.
}
