% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapSimilarity.R
\name{.ModelPvalue}
\alias{.ModelPvalue}
\title{Returns the variance array after permutations.}
\usage{
.ModelPvalue(
  dataSimulation,
  dataReference,
  clusterCut,
  permutations,
  refClusterVar,
  corMethod,
  simulatedClusterVar
)
}
\arguments{
\item{dataSimulation}{Simulation data to be compared.}

\item{dataReference}{Reference data with which to compare.}

\item{clusterCut}{The original cluster assignments.}

\item{permutations}{The number of permutations.}

\item{refClusterVar}{SD of the clusters.}

\item{corMethod}{Correlation method to be used.}

\item{simulatedClusterVar}{Variance of simulated clusters}
}
\value{
An array of dimension n.models by nClusters by permutations.
}
\description{
A utility function.
}
