% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapSimilarity.R
\name{.NthMin}
\alias{.NthMin}
\title{Find nth minimum value from a vector}
\usage{
.NthMin(x, index)
}
\arguments{
\item{x}{the given unsorted vector}

\item{index}{N.}
}
\value{
the nth minimum element of the vector
}
\description{
A utility function to find the nth minimum
}
