% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapSimilarity.R
\name{.PermutedVar}
\alias{.PermutedVar}
\title{Find variance of permutations}
\usage{
.PermutedVar(simulated.refCor, clusterCut, permutations, refClusterVar)
}
\arguments{
\item{simulated.refCor}{Correlation matrix of simulated and reference data}

\item{clusterCut}{The original cluster assignments}

\item{permutations}{The number of permutations}

\item{refClusterVar}{Reference Cluster Variance}
}
\value{
An array of dimension n.models by nClusters by permutations
}
\description{
A utility function to generate permutations
}
