% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapSimilarity.R
\name{.SimulatedPValueAbs}
\alias{.SimulatedPValueAbs}
\title{Finds the variance corresponding to a given value.}
\usage{
.SimulatedPValueAbs(permutedVar, simulatedClusterVar)
}
\arguments{
\item{permutedVar}{An array containing the distance of clusters for 
each model for every permutation.}

\item{simulatedClusterVar}{Variance of simulated clusters}
}
\value{
p-values for each model.
}
\description{
A utility function to calculate the absolute p values.
}
