% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMT1}
\alias{EMT1}
\title{A circuit for epithelial to mesenchymal transition}
\format{data frame}
\usage{
EMT1
}
\description{
A circuit for epithelial to mesenchymal transition
}
\keyword{datasets}
