% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMT2}
\alias{EMT2}
\title{A circuit for epithelial to mesenchymal transition including microRNAs}
\format{data frame}
\usage{
EMT2
}
\description{
A circuit for epithelial to mesenchymal transition including microRNAs
}
\keyword{datasets}
