% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{genParamNames}
\alias{genParamNames}
\title{Generate parameter names for a circuit}
\usage{
genParamNames(circuit = "inputs/test.tpo")
}
\arguments{
\item{circuit}{RacipeSE object or topology as data.frame or filename}
}
\value{
list
}
\description{
Generate parameter names for a circuit
}
\examples{
rSet <- RacipeSE()
data("demoCircuit")
sracipeCircuit(rSet) <- demoCircuit
paramNames <- sRACIPE::genParamNames(rSet)

}
