% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{ic<-}
\alias{ic<-}
\alias{ic<-,RacipeSE-method}
\alias{ic-set}
\title{A method to set the initial conditions}
\usage{
ic(.object) <- value

\S4method{ic}{RacipeSE}(.object) <- value
}
\arguments{
\item{.object}{RacipeSE object}

\item{value}{DataFrame containing the initial conditions}
}
\value{
A RacipeSE object
}
\description{
The circuit file should contain three columns with headers,
"Source" "Target" "Type"
Here "Source" and "Target" are the names of the genes and "Type" refers to
the regulation, "1" if source activates target and "2" if source inhibits
target.
}
\examples{
data("demoCircuit")
\dontrun{
rSet <- sRACIPE::simulateGRC(circuit = demoCircuit, numModels = 100)
ics <- ic(rSet)
ic(rSet) <- ics
}
}
