% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{ic}
\alias{ic}
\alias{ic,RacipeSE-method}
\title{A method to get the initial conditions used for simulations}
\usage{
ic(.object)

\S4method{ic}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
DataFrame
}
\description{
The initial conditions of each of the models.
}
\examples{
data("demoCircuit")
\dontrun{
rSet <- sRACIPE::simulateGRC(circuit = demoCircuit, numModels = 100)
ics <- ic(rSet)
ic(rSet) <- ics
}
}
