% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{load_data}
\alias{load_data}
\title{Load Data (Deprecated).}
\usage{
load_data(data_simulation = data_simulation, topology_df = topology)
}
\description{
Use \code{\link{simulateGRC}}

\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}
