% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTopology.R
\name{loadCircuit}
\alias{loadCircuit}
\title{Load the topology from a file.}
\usage{
loadCircuit(circuit = "inputs/test.tpo", header = TRUE)
}
\arguments{
\item{circuit}{\code{Character or data.frame}. Topology file address or as a
data.frame.}

\item{header}{(optional) \code{Logical}. Whether the input file has a header
or not.
\code{TRUE} by default.}
}
\value{
\code{List}. Contains the number of genes, circuit name,
topology (Source, Target, Type), and directory path.
}
\description{
Read an input file and return the circuit information
 as a \code{list}.
A typical topology file looks like
\tabular{lcr}{
  Source \tab Target \tab Type \cr
  geneA \tab geneB \tab 2 \cr
  geneB \tab geneC \tab 1 \cr
  geneB \tab geneA \tab 2
}
Here the regulation type is specified by number - activation: \code{1},
 inhibition: \code{2}
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}

\examples{
topologyInfo <- loadCircuit("inputs/test.tpo")

}
