% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{Load the sRACIPE configuration from a file.}
\usage{
loadConfig(configFile = "inputs/sRACIPE.cfg", header = TRUE)
}
\arguments{
\item{header}{(optional) Logical. Whether the input file has a header or not.
\code{TRUE} by default.}

\item{topologyFile}{Character. Topology file address.}
}
\value{
List. Contains various parameters and their values.
}
\description{
Read the sRACIPE configuration file to initialize the parameters.
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}

\examples{
\dontrun{
configuration <- loadConfig("inputs/sRACIPE.cfg")
}

}
