% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarity.R
\name{ModelPvalue}
\alias{ModelPvalue}
\title{Returns the variance array after permutations.}
\usage{
ModelPvalue(data.simulation, data.reference, cluster.cut, permutations,
  ref.cluster.var, cor.method, simulated.cluster.var)
}
\arguments{
\item{data.simulation}{Simulation data to be compared.}

\item{data.reference}{Reference data with which to compare.}

\item{cluster.cut}{The original cluster assignments.}

\item{permutations}{The number of permutations.}

\item{ref.cluster.var}{SD of the clusters.}
}
\value{
An array of dimension n.models by n.clusters by permutations.
}
\description{
A utility function.
}
