% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{normalizeGE}
\alias{normalizeGE}
\alias{normalizeGE,RacipeSE-method}
\title{Normalize the simulated gene expression}
\usage{
normalizeGE(.object)

\S4method{normalizeGE}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
A RacipeSE object
}
\description{
Normalize the simulated gene expression including gene
expressions for stochastic and knockout simulations
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockDownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
}

\examples{
data("demoCircuit")
\dontrun{
rSet <- sRACIPE::simulateGRC(circuit = demoCircuit, numModels = 100)
rSet <- normalizeGE(rSet)
}
}
