% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{params<-}
\alias{params<-}
\alias{params<-,RacipeSE-method}
\alias{params-set}
\title{A method to set the simulation parameters}
\usage{
params(.object) <- value

\S4method{params}{RacipeSE}(.object) <- value
}
\arguments{
\item{.object}{RacipeSE object}

\item{value}{DataFrame containing the parameteres}
}
\value{
A RacipeSE object
}
\description{
The circuit file should contain three columns with headers,
"Source" "Target" "Type"
Here "Source" and "Target" are the names of the genes and "Type" refers to
the regulation, "1" if source activates target and "2" if source inhibits
target.
}
\examples{
data("demoCircuit")
\dontrun{
rSet <- sRACIPE::simulateGRC(circuit = demoCircuit, numModels = 100)
parameters <- params(rSet)
params(rSet) <- parameters
}
}
