% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{params}
\alias{params}
\alias{params,RacipeSE-method}
\title{A method to access the simulation parameters}
\usage{
params(.object)

\S4method{params}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}

\item{...}{additional arguments}
}
\value{
A  data.frame
}
\description{
The parameters for each model.
}
\examples{
data("demoCircuit")
RacipeSet <- simulateGRC(demoCircuit)
parameters <- params(RacipeSet)
params(RacipeSet) <- parameters
}
