% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarity.R
\name{PermutedVar}
\alias{PermutedVar}
\title{Find variance of permutations}
\usage{
PermutedVar(simulated.ref.cor, cluster.cut, permutations, ref.cluster.var)
}
\arguments{
\item{simulated.ref.cor}{Correlation matrix of simulated and reference data}

\item{cluster.cut}{The original cluster assignments}

\item{permutations}{The number of permutations}
}
\value{
An array of dimension n.models by n.clusters by permutations
}
\description{
A utility function to generate permutations
}
