% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{plot_data_deterministic}
\alias{plot_data_deterministic}
\title{Plot the simulated data}
\usage{
plot_data_deterministic(data_file = output_file,
  plot_filename = plot_filename, topology = topology, bin_count = 40,
  plotTsne = FALSE)
}
\description{
Use \code{\link{plotData}}.

\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}
