% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{plot_data_knockout_all}
\alias{plot_data_knockout_all}
\title{Plot knockout data for all genes (Deprecated)}
\usage{
plot_data_knockout_all(output_file, output_file_KOpre, output_file_KOsuf,
  tfs_knockout, plot_filename = filename, topology_df = topology,
  config = configuration, bin_count = 40)
}
\description{
Use \code{\link{plotData}} instead.

\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}
