% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{plot_data_knockout_single}
\alias{plot_data_knockout_single}
\title{Plot knockout Data (Deprecated)}
\usage{
plot_data_knockout_single(output_file, output_file_knockout,
  plot_filename = filename, topology_df = topology,
  KNOCKOUT = NA_character_, config = configuration, bin_count = 40)
}
\description{
Use \code{\link{plotData}} instead.

\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}
