% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{plot_data_stochastic}
\alias{plot_data_stochastic}
\title{Plot Stochastic Data (Deprecated)}
\usage{
plot_data_stochastic(output_file, plot_filename = filename,
  topology_df = topology, config = configuration, bin_count = 40)
}
\description{
Use \code{\link{plotData}} instead.

\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}
