% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedFunctions.R
\name{plot_network}
\alias{plot_network}
\title{Plot the network (Deprecated)}
\usage{
plot_network(topology = topology, plotToFile = TRUE)
}
\description{
Use \code{\link{plotCircuit}}.

\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}
