% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{plotCircuit}
\alias{plotCircuit}
\alias{plotCircuit,RacipeSE-method}
\title{Plot Gene Regulatory Circuit}
\usage{
plotCircuit(.object, plotToFile = FALSE)

\S4method{plotCircuit}{RacipeSE}(.object, plotToFile = TRUE)
}
\arguments{
\item{.object}{RacipeSE object
A list returned by \code{\link{simulateGRC}} function}

\item{plotToFile}{(optional) logical. Default \code{FALSE}. Whether to save
plots to a file.}
}
\value{
circuit plot
}
\description{
Plot Gene Regulatory Circuit to a file or output device.
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockDownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}}
}

\examples{
data("demoCircuit")
rSet <- sRACIPE::simulateGRC(circuit = demoCircuit, numModels = 100)
plotCircuit(rSet, plotToFile = FALSE)
}
