% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RacipeSE-class}
\alias{RacipeSE-class}
\alias{.RacipeSE}
\title{RacipeSE}
\description{
An S4 class for Random Circuit Perturbation
simulations of networks. Extends the SummarizedExperiment class
}
\seealso{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} for
SummarizedExperiment documentation
}
