% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RacipeSE}
\alias{RacipeSE}
\title{RacipeSE object construct}
\usage{
RacipeSE(...)
}
\arguments{
\item{...}{Other arguments}
}
\value{
RacipeSE object
}
\description{
Create an RacipeSE object
}
\examples{
rSet <- RacipeSE()

}
