% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repressilator}
\alias{repressilator}
\title{A loop motif for demonstrating limit cycles in gene expression}
\format{
A data frame with 3 rows and 3 variables.
}
\usage{
repressilator
}
\description{
This data contains the topology for a repressilator circuit of
3 genes cyclically inhibiting each other with transcription factor
regulation. This is a simple circuit that can generate limit cycles in gene
expression in roughly 5% of models in sRACIPE.
}
\keyword{datasets}
