% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSimilarity.R
\name{SimulatedPValueAbs}
\alias{SimulatedPValueAbs}
\title{Finds the variance corresponding to a given value.}
\usage{
SimulatedPValueAbs(permuted.var, simulated.cluster.var)
}
\arguments{
\item{permuted.var}{An array containing the distance of clusters for each model for every permutation.}

\item{p.value}{Cut off p vlaue.}
}
\value{
p-values for each model.
}
\description{
A utility function to calculate the absolute p values.
}
