% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGRC.R
\name{simulateGRC}
\alias{simulateGRC}
\title{Simulate a gene regulatory circuit using RACIPE}
\usage{
simulateGRC(circuit = "inputs/test.tpo", config = "inputs/sRACIPE.cfg",
  anneal = FALSE, knockOut = NA_character_, numModels = 500,
  paramRange = 100, prodRateMin = 1, prodRateMax = 100,
  degRateMin = 0.1, degRateMax = 1, foldChangeMin = 1,
  foldChangeMax = 100, hillCoeffMin = 1L, hillCoeffMax = 6L,
  integrateStepSize = 0.02, simulationTime = 50, nIC = 1L,
  nNoise = 0L, initialNoise = 50, noiseScalingFactor = 0.5,
  shotNoise = 0, scaledNoise = FALSE, outputPrecision = 12L,
  printStart = 50, printInterval = 10, stepper = "RK4",
  thresholdModels = 10000, plots = FALSE, plotToFile = TRUE,
  genIC = TRUE, genParams = TRUE, integrate = TRUE,
  rkTolerance = 0.01)
}
\arguments{
\item{circuit}{List or character. The file containing the circuit or the list
returned by the function \code{\link{loadCircuit}}}

\item{config}{(optional) data.frame or character. The file containing the
configuration or the dataframe
returned by the function \code{\link{loadConfig}}}

\item{anneal}{(optional) Logical. Default FALSE. Whether to use annealing
for stochastic simulations. If TRUE, the gene expressions at higher noise
are used as initial conditions for simulations at lower noise.}

\item{knockOut}{(optional) List of character or vector of characters.
Simulation after knocking out one or more genes. To knock out all the genes
in the circuit, use \code{knockOut = "all"}. If it is a vector, then all
the genes in the vector will be knocked out simultaneously.}

\item{numModels}{(optional) Integer. Default 500. Number of random models
to be simulated.}

\item{paramRange}{(optional) numeric (0-100). Default 100. The relative
range of parameters (production rate, degradation rate, fold change).}

\item{prodRateMin}{(optional) numeric. Default 1. Minimum production rate.}

\item{prodRateMax}{(optional) numeric. Default 100. Maximum production rate.}

\item{degRateMin}{(optional) numeric. Default 0.1. Minimum degradation rate.}

\item{degRateMax}{(optional) numeric. Default 1. Maximum degradation rate.}

\item{foldChangeMin}{(optional) numeric. Default 1. Minimum fold change for
interactions.}

\item{foldChangeMax}{(optional) numeric. Default 100. Maximum fold change for
interactions.}

\item{hillCoeffMin}{(optional) integer. Default 1. Minimum hill coefficient.}

\item{hillCoeffMax}{(optional) integer. Default 6. Maximum hill coefficient.}

\item{integrateStepSize}{(optional) numeric. Default 0.02. step size for
integration using "EM" and "RK4" steppers.}

\item{simulationTime}{(optional) numeric. Total simulation time.}

\item{nIC}{(optional) integer. Default 1. Number of initial conditions to be
simulated for each model.}

\item{nNoise}{(optional) integer. Default 0.
Number of noise levels at which simulations
are to be done. Use nNoise = 1 if simulations are to be carried out at a
specific noise. If nNoise > 0, simulations will be carried out at nNoise
levels as well as for zero noise. "EM" stepper will be used for simulations
and any argument for stepper will be ignoired.}

\item{initialNoise}{(optional) numeric.
Default 50/sqrt(number of genes in the circuit). The initial value of noise
for simulations. The noise value will decrease by a factor
\code{noiseScalingFactor} at subsequent noise levels.}

\item{noiseScalingFactor}{(optional) numeric (0-1) Default 0.5.
The factor by which noise
will be decreased when nNoise > 1.}

\item{shotNoise}{(optional) numeric. Default 0.
The ratio of shot noise to additive
noise.}

\item{scaledNoise}{(optional) logical. Default FALSE. Whether to scale the
noise in each gene by its expected median expression across all models. If
TRUE the noise in each gene will be proportional to its expression levels.}

\item{printStart}{(optional) numeric (0-\code{simulationTime}).
 Default \code{simulationTime}
The time from which the output should be recorded. Useful for time series
analysis and studying the dynamics of a model for a particular initial
condition.}

\item{printInterval}{(optional) numeric (\code{integrateStepSize}-
\code{simulationTime - printStart}). Default 10. The separation between
two recorded time points.}

\item{stepper}{(optional) Character. Stepper to be used for integrating the
differential equations. The options include \code{"EM"} for Euler-Maruyama
O(1), \code{"RK4"}
for fourth order Runge-Kutta O(4) and \code{"DP"} for adaptive stepper based
Dormand-Prince algorithm. The default method is \code{"RK4"} for deterministic
simulations and the method defaults to \code{"EM"} for stochastic simulations.}

\item{thresholdModels}{(optional)  integer. Default 10000. The number of
models to be used for calculating the thresholds for genes.}

\item{plots}{(optional) logical Default \code{TRUE}.
Whether to plot the simuated data.}

\item{plotToFile}{(optional) Default \code{FALSE}. Whether to save the plots
to a file.}

\item{genIC}{(optional) logical. Default \code{TRUE}. Whether to generate
the initial conditions. If \code{FALSE}, the initial conditions must be
supplied as a dataframe to \code{circuit$ic}.}

\item{genParams}{(optional) logical. Default \code{TRUE}. Whether to generate
the parameters. If \code{FALSE}, the parameters must be
supplied as a dataframe to \code{circuit$params}.}

\item{integrate}{(optional) logical. Default \code{TRUE}. Whether to
integrate the differential equations or not. If \code{FALSE}, the function
will only generate the parameters and initial conditions. This can be used
iteratively as one can fist generate the parameters and initial conditions
and then modify these before using these modified values for integration.
For example, this can be used to knockOut genes by changing the production
rate and initial condition to zero.}

\item{rkTolerance}{(optional) numeric. Default \code{0.01}. Error tolerance
for adaptive integration method.}

\item{outputPrecison}{(optional) integer. Default 12.
The decimal point precison of
the output to be printed in the tmp folder.}
}
\value{
List. A list containing the circuit, parameters, initial conditions,
simulated gene expressions, and simulation configuration.
}
\description{
Simulate a gene regulatory circuit using its topology as the
only input. It will generate an ensemble of random models.
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}

