% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{sRACIPE_load_configuration}
\alias{sRACIPE_load_configuration}
\title{(deprecated) Load the sRACIPE configuration from a file.}
\usage{
sRACIPE_load_configuration(config_file = "inputs/sRACIPE.cfg")
}
\arguments{
\item{topologyFile}{Character. Topology file address.}

\item{header}{(optional) Logical. Whether the input file has a header or not.
\code{TRUE} by default.}
}
\value{
List. Contains various parameters and their values.
}
\description{
Read the sRACIPE configuration file to initialize the parameters.
Use the function \code{\link{loadConfig}} instead.
}
\examples{
\dontrun{
configuration <- loadConfig("inputs/sRACIPE.cfg")
}

}
