% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTopology.R
\name{sRACIPE_load_topology}
\alias{sRACIPE_load_topology}
\title{(deprecated) Load the topology from a file.}
\usage{
sRACIPE_load_topology(topology_file = "inputs/test.tpo")
}
\arguments{
\item{topology_file}{Character. Topology file address.}
}
\value{
List. Contains the number of genes, circuit name,
topology (Source, Target, Type), and directory path.
}
\description{
Read an input file and return the circuit information as a list.
This function will be removed in future versions.
Use the function \code{\link{loadTopology}} instead.
}
