% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{sracipeConfig}
\alias{sracipeConfig}
\alias{sracipeConfig,RacipeSE-method}
\title{A method to access the simulation hyperparameters}
\usage{
sracipeConfig(.object)

\S4method{sracipeConfig}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
list
}
\description{
The hyperparameters like number of models, range from which
parameters are to be sampled, simulation time etc.
}
\examples{
RacipeSet <- RacipeSE()
data("demoCircuit")
sracipeCircuit(RacipeSet) <- demoCircuit
sracipeConfig(RacipeSet)
rm(RacipeSet)
}
