% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{sracipeConverge}
\alias{sracipeConverge}
\alias{sracipeConverge,RacipeSE-method}
\title{A method to get the convergence results for deterministic simulations}
\usage{
sracipeConverge(.object)

\S4method{sracipeConverge}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
DataFrame
}
\description{
Gathers the convergence and speed of convergence for each model
and condition for deterministic simulations
}
\examples{
data("demoCircuit")
rSet <- sRACIPE::sracipeSimulate(circuit = demoCircuit, numModels = 20)
cd <- sracipeIC(rSet)
rm(rSet, cd)
}
