% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{sracipeConvergeDist}
\alias{sracipeConvergeDist}
\alias{sracipeConvergeDist,RacipeSE-method}
\title{A method to visualize convergence distributions}
\usage{
sracipeConvergeDist(.object, plotToFile = FALSE)

\S4method{sracipeConvergeDist}{RacipeSE}(.object, plotToFile = FALSE)
}
\arguments{
\item{.object}{RacipeSE object generated by \code{\link{sracipeSimulate}}
function.}

\item{plotToFile}{(optional) logical. Default \code{FALSE}. Whether to save
plots to a file.}
}
\value{
\code{RacipeSE} object
}
\description{
When convergence tests are done (deterministic systems with time
series off), this method creates a plot of the proportion of converged initial
conditions as the number of convergence tests increases, up to the total number
done by the simulation
}
\section{Related Functions}{


\code{\link{sracipeSimulate}},  \code{\link{sracipeKnockDown}},
\code{\link{sracipeOverExp}},  \code{\link{sracipePlotData}}
}

\examples{
data("demoCircuit")
\dontrun{
rSet <- sRACIPE::sracipeSimulate(circuit = demoCircuit, numModels = 20,
integrateStepSize = 0.1, numConvergenceTests = 30)
rSet <- sracipeConvergeDist(rSet)
}
}
