% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sracipeGenParamNames}
\alias{sracipeGenParamNames}
\title{Generate parameter names for a circuit}
\usage{
sracipeGenParamNames(circuit = "inputs/test.tpo")
}
\arguments{
\item{circuit}{RacipeSE object or topology as data.frame or filename}
}
\value{
list
}
\description{
Generate parameter names for a circuit
}
\examples{
rSet <- RacipeSE()
data("demoCircuit")
sracipeCircuit(rSet) <- demoCircuit
paramNames <- sRACIPE::sracipeGenParamNames(rSet)

}
