% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{sracipeGetTS}
\alias{sracipeGetTS}
\alias{sracipeGetTS,RacipeSE-method}
\title{A method to extract the time series}
\usage{
sracipeGetTS(.object)

\S4method{sracipeGetTS}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
List
}
\description{
If timeSeries option is used in sracipeSimulate function, this
method will return the simulated time series.
}
\examples{
data("demoCircuit")
RacipeSet <- RacipeSE()
sracipeCircuit(RacipeSet) <- demoCircuit
RacipeSet <- sracipeSimulate(demoCircuit, timeSeries = TRUE, 
simulationTime = 2)
trajectories <- sracipeGetTS(RacipeSet)
rm(RacipeSet)
}
