% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{sracipeIC}
\alias{sracipeIC}
\alias{sracipeIC,RacipeSE-method}
\title{A method to get the initial conditions used for simulations}
\usage{
sracipeIC(.object)

\S4method{sracipeIC}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
DataFrame
}
\description{
The initial conditions of each of the models.
}
\examples{
data("demoCircuit")
rSet <- sRACIPE::sracipeSimulate(circuit = demoCircuit, numModels = 20,
integrate=FALSE)
ics <- sracipeIC(rSet)
rm(rSet,ics)
}
