% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{sracipeParams<-}
\alias{sracipeParams<-}
\alias{sracipeParams<-,RacipeSE-method}
\alias{sracipeParams-set}
\title{A method to set the simulation parameters}
\usage{
sracipeParams(.object) <- value

\S4method{sracipeParams}{RacipeSE}(.object) <- value
}
\arguments{
\item{.object}{RacipeSE object}

\item{value}{DataFrame containing the parameteres}
}
\value{
A RacipeSE object
}
\description{
Set the parameters
}
\examples{
data("demoCircuit")
rSet <- sRACIPE::sracipeSimulate(circuit = demoCircuit, numModels = 20, 
integrate = FALSE)
parameters <- sracipeParams(rSet)
sracipeParams(rSet) <- parameters
rm(parameters, rSet)
}
