% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{sracipeParams}
\alias{sracipeParams}
\alias{sracipeParams,RacipeSE-method}
\title{A method to access the simulation parameters}
\usage{
sracipeParams(.object)

\S4method{sracipeParams}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
A  data.frame
}
\description{
The parameters for each model.
}
\examples{
data("demoCircuit")
RacipeSet <- sracipeSimulate(demoCircuit, integrate = FALSE, numModels=20)
parameters <- sracipeParams(RacipeSet)
sracipeParams(RacipeSet) <- parameters
rm(parameters,RacipeSet)
}
