% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToggleSwitchSA}
\alias{ToggleSwitchSA}
\title{A toggle switch circuit with self activations}
\format{data frame}
\usage{
ToggleSwitchSA
}
\description{
A toggle switch circuit with self activations
}
\keyword{datasets}
