% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellCycle}
\alias{cellCycle}
\title{A circuit for modeling the yeast cell cycle}
\format{
A data frame with 38 rows and 3 variables.
}
\usage{
cellCycle
}
\description{
This data contains a circuit with 15 nodes representing genes
involved with the yeast cell cycle.
For further details,
see Katebi A, Kohar V, Lu M. Random Parametric Perturbations of Gene
Regulatory Circuit Uncover State Transitions in Cell Cycle. iScience. 2020
Jun 26;23(6):101150. doi: 10.1016/j.isci.2020.101150. Epub 2020 May 11.
PMID: 32450514; PMCID: PMC7251928.
}
\keyword{datasets}
