% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoupledToggleSwitchSA}
\alias{CoupledToggleSwitchSA}
\title{Five coupled toggle switches}
\format{A data frame with 18 rows and 3 variables}
\usage{
CoupledToggleSwitchSA
}
\description{
This data contains the topology of a 
circuit with ten genes A1...A5 and B1...B5. 
Genes with different alphabet inhibit each other 
whereas genes from different groups activate the other.
For further details,
see Kohar, V. and Lu, M., Role of noise and parametric 
variation in the dynamics of gene regulatory circuits, 
npj Systems Biology and Applications 4, Article number: 40 (2018)
}
\keyword{datasets}
