% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demoCircuit}
\alias{demoCircuit}
\title{A toggle switch circuit for demonstrations}
\format{
A data frame with 4 rows and 3 variables
}
\usage{
demoCircuit
}
\description{
This data contains the topology of a 
circuit with two genes A and B both of which inhibit each
other and have self activations. For further details,
see Kohar, V. and Lu, M., Role of noise and parametric 
variation in the dynamics of gene regulatory circuits, 
npj Systems Biology and Applications 4, Article number: 40 (2018)
}
\keyword{datasets}
