% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{densityPlot}
\alias{densityPlot}
\title{Density Plot}
\usage{
densityPlot(plotData, binCount = 40, plotColor = NULL)
}
\arguments{
\item{plotData}{Dataframe containing the data.}

\item{binCount}{(optional) Integer. Default 40. The number of bins to be used for
dividing the data along an axis.}

\item{plotColor}{(optional) The color palette.}
}
\description{
Plot the density of points as an image alongwith histograms on
the sides.
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}

