% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{knockdownAnalysis}
\alias{knockdownAnalysis}
\title{Perform in-silico knockdown analysis}
\usage{
knockdownAnalysis(rSet = NULL, topologyDf = NULL, dataFile = NULL,
  paramFile = NULL, reduceProduction = 30, nClusters = 2,
  clusterOfInterest = 2, plotFilename = plot_filename,
  plotHeatmap = TRUE, plotBarPlot = TRUE, clusterCut = cluster_cut,
  plotToFile = TRUE)
}
\arguments{
\item{rSet}{(optional) List. A list returned by \code{\link{simulateGRC}}
function.}

\item{topologyDf}{(optional) Deprecated.
Topology (list) generated by sRACIPE_load_topology function}

\item{dataFile}{(optional) Deprecated. Gene expression file (character).}

\item{paramFile}{(optional) Deprecated. Parameter file (character).}

\item{reduceProduction}{(optional) Percentage to which production rate
decreases on knockdown. Uses a default value of 10 percent.}

\item{nClusters}{(optional) Number of clusters in the data. Uses a default
value of 2.}

\item{clusterOfInterest}{(optional) cluster number (integer) to be used for arranging
the transcription factors}

\item{plotFilename}{(optional) Name of the output file.
Uses the topology name as default.}

\item{cluster}{assignments of the models}
}
\value{
List containing fraction of models in different clusters
 in the original simulations and after knowcking down different genes.
Additionaly, it generates two pdf files in the results folder. First is barplot
showing the percentage of different clusters in the original simulations
and after knocking down each gene. The second pdf contains the heatmap of
clusters after marking the models with cluster assignments.
}
\description{
Calculate the fraction of models in different clusters with full parameter
range and on a subset of models with low production rate of a specific gene
 representing the knockdown of the specific gene.
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}

