% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{normalizeGE}
\alias{normalizeGE}
\title{Normalize the gene expression data}
\usage{
normalizeGE(rSet = rSet)
}
\arguments{
\item{rSet}{List. List returned by the function \code{\link{simulateGRC}}
containing the circuit, simulated gene expressions etc.}
}
\description{
Normalize the sRACIPE simulated gene expression data
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}

