% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCircuit}
\alias{plotCircuit}
\title{Plot Gene Regulatory Circuit}
\usage{
plotCircuit(rSet = rSet, plotToFile = TRUE)
}
\arguments{
\item{rSet}{List A list returned by \code{\link{simulateGRC}} function}

\item{plotToFile}{(optional) logical. Default \code{TRUE}. Whether to save
plots to a file.}
}
\description{
Plot Gene Regulatory Circuit to a file or output device.
}
\section{Related Functions}{


\code{\link{simulateGRC}},  \code{\link{knockdownAnalysis}},
\code{\link{overExprAnalysis}},  \code{\link{plotData}},
\code{\link{calcHeatmapSimilarity}}
}

