% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotParamBifur}
\alias{plotParamBifur}
\title{Parameter bifurcation plots}
\usage{
plotParamBifur(rSet = rSet, paramName, data = NULL,
  paramValue = NULL, assignedClusters = NULL, plotToFile = TRUE)
}
\arguments{
\item{rSet}{List. The list generated by \code{\link{simulateGRC}} function.}

\item{paramName}{character. The name of the parameter to be plotted.}

\item{data}{(optional) dataframe. Default rSet geneExpression. The data
to be plotted. For example, use rSet$stochasticSimulations$[noise] to plot
the stochastic data.}

\item{paramValue}{(optional) Dataframe. The parameter values if rSet$params
is not to be used.}

\item{assignedClusters}{(optional) Dataframe. The cluster assignment of data.}
}
\description{
Plot the expression of the genes against parameter values
to understand the effect of parameters on the gene expressions.
}
