% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RacipeSE-class}
\alias{RacipeSE-class}
\alias{.RacipeSE}
\title{RacipeSE}
\description{
An S4 class for Random Circuit Perturbation (RACIPE)
simulations of networks. Extends the \link{SummarizedExperiment} class.
RACIPE can simulate a gene regulatory circuit using the circuit and
a large ensemble of parameters.
}
