% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{sracipeCombineRacipeSE}
\alias{sracipeCombineRacipeSE}
\alias{sracipeCombineRacipeSE,list-method}
\title{A method to combine RacipeSE objects with the same config}
\usage{
sracipeCombineRacipeSE(.object)

\S4method{sracipeCombineRacipeSE}{list}(.object)
}
\arguments{
\item{.object}{RacipeSE object generated by \code{\link{sracipeSimulate}}
function.}
}
\value{
\code{RacipeSE} object
}
\description{
It is often the case that to save time, multiple simulations
of the same topology with the same config file are run in parallel. However,
one still wants to condense the results of the different simulations into one
object. This method does that, combining the params, ics, and expressions
of a list of objects into one.
}
\section{Related Functions}{


\code{\link{sracipeSimulate}},  \code{\link{sracipeKnockDown}},
\code{\link{sracipeOverExp}},  \code{\link{sracipePlotData}}
}

\examples{
data("demoCircuit")
\dontrun{
rSet <- sRACIPE::sracipeSimulate(circuit = demoCircuit, numModels = 20,
integrateStepSize = 0.1, numConvergenceTests = 30)
rSet <- sracipeConvergeDist(rSet)
}
}
